backend/requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# PDF processing
PyMuPDF==1.23.8

# AI and LLM dependencies
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10

# Vector database and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Text processing
tiktoken==0.5.2

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Environment and configuration
python-dotenv==1.0.0

# TTS Podcast Dependencies
openai>=1.40.0

# Additional utilities
requests==2.31.0
aiofiles==23.2.1